requirements.txt
Django==4.0.5
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
PyJWT==2.4.0
Pillow==9.1.1
django-rest-swagger==2.2.0
django-filter==22.1